moved pub args class
